+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - when doing a text search in your note the selected text is now preset as 
   search text if there is any and there is no other search text filled in
 - added more information to the debug tab in the settings
+- the todo tab in the settings is now disabled if there are no ownCloud settings
+  (for [Issue #208](https://github.com/pbek/QOwnNotes/issues/208))
 
 ## 16.06.6
 - you now can use a real **dark mode** on every operating system